backend+ai/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.3.0
flask-jwt-extended==4.5.3
gunicorn==21.2.0
PyJWT==2.8.0
cryptography==41.0.7